Create HdfsMrsImageReaderBuilder and MapFileReaderBuilder
Created a builder for HdfsMrsImageReader mocks and for MapFile.Reader mocks.  Due to similarities in the interface for SequenceFile.Reader and MapFile.Reader, factored out a KeyValueHelper and delegated the key/value related functions to that class.  Refactored SequenceFile.Reader to use the KeyValueHelper class.  Added myself as a developer to the main pom.xml

Signed-off-by: ericwood73 <ericwood73@gmail.com>